Split watched files between assets and templates
