+++ Gruntfile.js
@@ -5,37 +5,29 @@ module.exports = function(grunt) {
 
     // Configuration of 'grunt-contrib-clean' task, to remove all output folder
     clean: {
-      build: ['output/'],
+      output: ['output/'],
+      assets: ['output/assets/'],
+      templates: ['output/templates/', 'output/locales', 'output/*.html'],
       dist: ['*.jar']
     },
 
     // Configuration of 'grunt-contrib-copy' task, to move files into the output folder
     copy: {
-      dist: {
+      assets: {
         files: [
           { expand: true, cwd: 'input/', dest: 'output/', src: 'assets/css/*.css' },
           { expand: true, cwd: 'input/', dest: 'output/', src: 'assets/js/*.js' },
           { expand: true, cwd: 'input/', dest: 'output/', src: 'assets/img/**/*' },
-          { expand: true, cwd: 'input/', dest: 'output/', src: 'assets/fonts/**/*' },
-          { expand: true, cwd: 'input/', dest: 'output/', src: '*.html' },
+          { expand: true, cwd: 'input/', dest: 'output/', src: 'assets/fonts/**/*' }
+        ]
+      },
+      templates: {
+        files: [
           { expand: true, cwd: 'input/', dest: 'output/', src: 'templates/*.hbs' },
           { expand: true, cwd: 'locales/', dest: 'output/locales', src: '**/*.yaml' },
           { expand: true, cwd: 'input/templates/partials/', dest: 'output/templates/', src: '**/*.json' },
           { expand: true, cwd: 'input/templates/partials/', dest: 'output/templates/', src: '**/*.hbs' }
         ]
-      }
-    },
-
-    // Configuration of 'grunt-contrib-coffee' task, to compile Coffeescript files into Javascript
-    coffee: {
-      dist: {
-        options: {
-          // fails when no coffee file found
-          //join: true
-        },
-        files: {
-          'output/assets/js/coffee.js': 'input/assets/js/*.coffee'
-        }
       }
     },
 
@@ -79,14 +71,41 @@ module.exports = function(grunt) {
       }
     },
 
+    // Configuration of the 'i18next' task, to support internationalization in Handlebars
+    i18next: {
+      options: {
+        preload: ['de', 'en'],
+        lng: 'en',
+        fallbackLng: 'en',
+        ns: {
+          namespaces: ['translations', 'home', 'catalog', 'checkout', 'my-account-login', 'no-search-result', 'mix-match', 'my-account'],
+          defaultNs: 'translations'
+        }
+      }
+    },
+
+    // Configuration of 'json-refs' task, to resolve JSON references
+    'json-refs': {
+      dist: {
+        files: [{
+          expand: true,
+          cwd: 'input/templates',
+          src: '*.json',
+          dest: 'output/templates',
+          ext: '.json'
+        }]
+      }
+    },
+
     // Configuration of 'grunt-contrib-watch' task, to watch for changes in order to run the build task again
     watch: {
-      scripts: {
-        files: [
-          'input/**/*',
-          'locales/**/*'
-        ],
-        tasks: ['build']
+      assets: {
+        files: ['input/assets/**/*'],
+        tasks: ['build-assets']
+      },
+      templates: {
+        files: ['input/templates/**/*', 'locales/**/*'],
+        tasks: ['build-templates']
       }
     },
 
@@ -98,8 +117,7 @@ module.exports = function(grunt) {
         artifactId: "<%= pkg.name %>",
         version: "<%= pkg.version %>",
         destFolder: "/META-INF/resources/webjars",
-        gitpush: true,
-        mode: "patch"
+        gitpush: true
       },
       install : {
         options : {
@@ -138,50 +156,25 @@ module.exports = function(grunt) {
         base: 'output'
       },
       src: ['**/*']
-    },
-
-    // Configuration of the 'i18next' task, to support internationalization in Handlebars
-    i18next: {
-      options: {
-        preload: ['de', 'en'],
-        lng: 'en',
-        fallbackLng: 'en',
-        ns: {
-          namespaces: ['translations', 'home', 'catalog', 'checkout', 'my-account-login', 'no-search-result', 'mix-match', 'my-account'],
-          defaultNs: 'translations'
-        }
-      }
-    },
-
-    // Configuration of 'json-refs' task, to resolve JSON references
-    'json-refs': {
-      dist: {
-        files: [{
-          expand: true,
-          cwd: 'input/templates',
-          src: '*.json',
-          dest: 'output/templates',
-          ext: '.json'
-        }]
-      }
     }
   });
 
+  grunt.loadTasks('tasks');
   grunt.loadNpmTasks('grunt-contrib-clean');
   grunt.loadNpmTasks('grunt-contrib-copy');
   grunt.loadNpmTasks('grunt-contrib-sass');
   grunt.loadNpmTasks('grunt-postcss');
-  grunt.loadNpmTasks('grunt-contrib-coffee');
   grunt.loadNpmTasks('grunt-compile-handlebars');
   grunt.loadNpmTasks('grunt-contrib-watch');
   grunt.loadNpmTasks('grunt-maven-tasks');
   grunt.loadNpmTasks('grunt-gh-pages');
 
   grunt.registerTask('default', ['build', 'watch']);
-  grunt.registerTask('build', ['clean', 'copy:dist', 'coffee', 'sass', 'postcss', 'pre-handlebars', 'handlebars']);
-  grunt.registerTask('release-patch', ['build', 'maven', 'clean:dist']);
-  grunt.registerTask('release-minor', ['build', 'maven:release:minor', 'clean:dist']);
-  grunt.registerTask('release-major', ['build', 'maven:release:major', 'clean:dist']);
+  grunt.registerTask('build', ['clean:dist', 'build-assets', 'build-templates']);
+  grunt.registerTask('build-assets', ['clean:assets', 'copy:assets', 'sass', 'postcss']);
+  grunt.registerTask('build-templates', ['clean:templates', 'copy:templates', 'pre-handlebars', 'handlebars']);
+  grunt.registerTask('release', ['build', 'maven:release', 'clean:dist']);
+  grunt.registerTask('install', ['build', 'maven:install', 'clean:dist']);
   grunt.registerTask('publish', ['gh-pages-clean', 'build', 'gh-pages']);
 
   grunt.registerTask('pre-handlebars', 'Tasks to be run before Handlebars', function() {
@@ -193,111 +186,4 @@ module.exports = function(grunt) {
     grunt.task.requires('pre-handlebars');
     grunt.task.run('compile-handlebars');
   });
-
-  grunt.registerTask('i18next', 'Internationalization init', function() {
-    var done = this.async();
-    var options = this.options({
-      preload: ['en'],
-      lng: 'en',
-      fallbackLng: 'en',
-      getAsync: false,
-      debug: false,
-      ns: {
-        namespaces: ['translations'],
-        defaultNs: 'translations'
-      },
-      resGetPath: 'locales/__lng__/__ns__.yaml'
-    });
-    Handlebars = require('handlebars');
-
-    i18n = require('i18next');
-
-    var yamlSync = require('i18next.yaml');
-    i18n.backend(yamlSync);
-
-    i18n.init(options, function (err, t) {
-      done(true);
-    });
-  });
-
-  grunt.registerMultiTask('json-refs', 'Resolves all JSON References and returns a fully resolved equivalent', function() {
-    var done = this.async();
-    // Default task configuration
-    var options = this.options({
-      partials: "input/templates/partials/"
-    });
-    // 'json-refs' configuration
-    var jsonRefsOptions = {
-      location: options.partials
-    };
-
-    var jsonRefs = require('json-refs');
-    var path = require('path');
-
-    var resolvedRefsPromises = this.files
-    .filter(removeInvalidFiles)
-    .filter(removeInexistentFiles)
-    .map(function(file) {
-      // Resolve JSON references
-      var json = grunt.file.readJSON(file.src[0]);
-      return jsonRefs.resolveRefs(json, jsonRefsOptions)
-      .then(function(result) {
-        return writeResolvedFile(file, result);
-      });
-    });
-
-    Promise.all(resolvedRefsPromises)
-    .then(function(result) {
-      done(true);
-    });
-  });
-
-  var removeInvalidFiles = function(file) {
-    if (file.src.length != 1) {
-      grunt.fail.warn("Only a single source file is currently supported.");
-      return false;
-    } else {
-      return true;
-    }
-  };
-
-  var removeInexistentFiles = function(file) {
-    var filepath = file.src[0];
-    if(!grunt.file.exists(filepath)) {
-      grunt.log.warn('Source file "' + filepath + '" not found.');
-      return false;
-    } else {
-      return true;
-    }
-  };
-
-  var writeResolvedFile = function(file, result) {
-    var written = false;
-    parseMetadata(result.metadata);
-    // Write the resolved JSON to a new file
-    written = grunt.file.write(file.dest, stringifyJson(result.resolved, 2));
-    if (written) {
-      grunt.log.debug('File "' + file.dest + '" created');
-    } else {
-      grunt.log.error('File "' + file.dest + '" failed on creation');
-    }
-    return written;
-  };
-
-  var parseMetadata = function(json) {
-    for (var key in json) {
-      if (json.hasOwnProperty(key)) {
-        var value = json[key];
-        if (value.hasOwnProperty("err")) {
-          grunt.log.error(stringifyJson(value, 2));
-        } else {
-          grunt.verbose.writeln(stringifyJson(value, 0));
-        }
-      }
-    }
-  }
-
-  var stringifyJson = function(json, space) {
-    return JSON.stringify(json, null, space);
-  }
 };
